Refactored the deleteConfirmed method in the Product controller to check for nulls and redirect to the index once done

+++ CPW219-AspnetMVC-CRUD-Debugging/Controllers/ProductsController.cs
@@ -74,8 +74,17 @@ namespace CPW219_AspnetMVC_CRUD_Debugging.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var product = await _context.Product.FindAsync(id);
-            _context.Product.Remove(product);
-            return RedirectToAction(nameof(Index));
+
+            if (product != null)
+            {
+                _context.Product.Remove(product);
+                await _context.SaveChangesAsync();
+                // Add a message here after debugging the project
+                return RedirectToAction(nameof(Index));
+            }
+
+            // Add a message here after debugging the project
+            return RedirectToAction("Index");
         }
 
         private bool ProductExists(int id)
